requirements.txt
openai-whisper
ffmpeg-python
sumy
pydub